+++ src/main/java/org/toilelibre/libe/soundtransform/model/library/note/Silence.java
@@ -22,7 +22,7 @@ public class Silence implements Note {
     }
 
     @Override
-    public int getFrequency () {
+    public float getFrequency () {
         return 0;
     }
