+++ src/test/java/org/toilelibre/libe/soundtransform/ShapeTest.java
@@ -41,7 +41,7 @@ public class ShapeTest extends SoundTransformTest {
 
         $.create (ConvertAudioFileService.class).writeInputStream (outputStream, output);
 
-        final int frequency = $.create (Sound2NoteService.class).convert ("output chord_note", $.create (TransformSoundService.class, new Slf4jObserver ()).fromInputStream ($.create (ConvertAudioFileService.class).callConverter (output))).getFrequency ();
+        final float frequency = $.create (Sound2NoteService.class).convert ("output chord_note", $.create (TransformSoundService.class, new Slf4jObserver ()).fromInputStream ($.create (ConvertAudioFileService.class).callConverter (output))).getFrequency ();
         new Slf4jObserver ().notify ("Output chord note should be around 387Hz, but is " + frequency + "Hz");
 
     }
@@ -58,7 +58,7 @@ public class ShapeTest extends SoundTransformTest {
 
         $.create (ConvertAudioFileService.class).writeInputStream (outputStream, output);
 
-        final int frequency = $.create (Sound2NoteService.class).convert ("output chord_note", $.create (TransformSoundService.class, new Slf4jObserver ()).fromInputStream ($.create (ConvertAudioFileService.class).callConverter (output))).getFrequency ();
+        final float frequency = $.create (Sound2NoteService.class).convert ("output chord_note", $.create (TransformSoundService.class, new Slf4jObserver ()).fromInputStream ($.create (ConvertAudioFileService.class).callConverter (output))).getFrequency ();
         new Slf4jObserver ().notify ("Output chord note should be around 332Hz, but is " + frequency + "Hz");
 
     }
